refactor: #538 내 토론방 UiState 분리
